backend/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
factory-boy==3.2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
